Wrap top-level server-side RPC call in RevokerMembrane
Capabilities can be called past the end of an RPC session if that session is ended
by the IoContext getting aborted. This doesn't result in any JS getting executed
since the call to IoContext::run immediately returns the abort exception, but this
might lead to calls attempting to schedule work after the IncomingRequest is getting
torn down. This change updates ServerTopLevelMembrane to wrap the top-level capability
in a RevokerMembrane on top of the existing CompletionMembrane.